Time: 17 ms (38.23%) | Memory: 109.6 MB (45.06%) - LeetSync

+++ 2352-design-an-atm-machine/design-an-atm-machine.cpp
@@ -4,7 +4,7 @@ public:
     vector<int> denomination;
     ATM() {
         notes=vector<long long>(5,0);
-        denomination={20,50,100,200,500};
+        denomination={20,50,100,200,500}; 
     }
     void deposit(vector<int> banknotesCount) {
         for(int i=0;i<5;i++){
@@ -13,19 +13,19 @@ public:
     }
     vector<int> withdraw(int amount) {
         vector<int> used(5,0);
-            int remain=amount;
-            for(int i=4;i>=0;i--){
-                long long canTake=min(notes[i],(long long)remain/denomination[i]);
-                used[i]=canTake;
-                remain-=canTake*denomination[i];
-            }
-            if(remain!=0){
-                return {-1};
-            }
-            for(int i=0;i<5;i++){
-                notes[i]-=used[i];
-            }
-            return used;
+        int remain=amount;
+        for(int i=4;i>=0;i--){
+            int canTake=min(notes[i],(long long)remain/denomination[i]);
+            used[i]=canTake;
+            remain-=canTake*denomination[i];
+        }
+        if(remain!=0){
+            return {-1};
+        }
+        for(int i=0;i<5;i++){
+            notes[i]-=used[i];
+        }
+        return used;
     }
 };
